Refactored: code for float value

+++ Max_of_three.java
@@ -2,8 +2,8 @@ package maxOfThree;
 
 public class Max_of_three {
 	
-	public static Integer maximum(Integer val1, Integer val2, Integer val3) {
-		int max= val1;
+	public static Float maximum(Float val1, Float val2, Float val3) {
+		Float max = val1;
 		if(val2.compareTo(max)>0) {
 			max=val2;
 		}else if(val3.compareTo(max)>0) {
@@ -12,7 +12,7 @@ public class Max_of_three {
 		return max;
 	}
 	public static void main(String[] args) {
-		Integer val1=20, val2=40, val3=30;
+		Float val1=(float) 20.0, val2=(float) 40.0, val3=(float)30.0;
 		System.out.println("Maximun number is "+maximum(val1,val2,val3));
 
 	}
